imagenet/requirements.txt
antialiased-cnns==0.3
appdirs==1.4.4
blessed @ file:///croot/blessed_1701887495364/work
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
certifi @ file:///croot/certifi_1707229174982/work/certifi
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
docker-pycreds==0.4.0
## !! Could not determine repository location
-e /root/miniconda/envs/lgd/lib/python3.8/site-packages
filelock @ file:///croot/filelock_1700591183607/work
gitdb==4.0.11
GitPython==3.1.43
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455532332/work
gpustat @ file:///croot/gpustat_1701888826609/work
idna @ file:///croot/idna_1666125576474/work
Jinja2 @ file:///croot/jinja2_1706733616596/work
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
networkx @ file:///croot/networkx_1690561992265/work
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
nvidia-ml-py @ file:///croot/nvidia-ml-py_1701875136273/work
pillow @ file:///croot/pillow_1707233021655/work
protobuf==4.25.3
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
PyYAML==6.0.1
requests @ file:///croot/requests_1707355572290/work
sentry-sdk==1.45.0
setproctitle==1.3.3
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.1
sympy @ file:///croot/sympy_1701397643339/work
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
triton==2.0.0
typing_extensions @ file:///croot/typing_extensions_1705599297034/work
urllib3 @ file:///croot/urllib3_1707770551213/work
wandb==0.16.6
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
